tests refactoring, add table_name parameter to methods
